Fixed a g++ warning
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@13021 01821946-851b-0410-b7f4-f1477c70e57a

+++ 3d/buildmodel.cpp
@@ -378,7 +378,7 @@ if (ogl.m_features.bVertexBufferObjects) {
 	}
 m_sortedVerts.Destroy ();
 
-for (i = 0; i < m_textures.Length (); i++)
+for (i = 0; i < (int) m_textures.Length (); i++)
 	m_textures [i].PrepareTexture (1, 0);
 }
